+++ src/app/page.tsx
@@ -2,7 +2,7 @@ import Hero         from '@/components/sections/Hero/Hero';
 import About        from '@/components/sections/About/About';
 import Projects     from '@/components/sections/Projects/Projects';
 import Skills       from '@/components/sections/Skills/Skills';
-import Certificates from '@/components/sections/Certificates/Certificates';
+import Certificates from '@/components/sections/Certificates';
 import Contact      from '@/components/sections/Contact/Contact';
 import Footer       from '@/components/common/Footer/Footer';